construct unique_operator_id consistently
